dezing, denoise tests updated
