+++ savu/test/travis/plugin_tests/filter_tests/denoise/ccpi_denoising_cpu_test.py
@@ -32,16 +32,16 @@ import savu.test.base_checkpoint_test
 import tempfile
 import os
 
-class CcpiDenoisingGpuTest(unittest.TestCase):
+class CcpiDenoisingCpuTest(unittest.TestCase):
 
-    def test_ccpi_denoise_gpu(self):
+    def test_ccpi_denoise_cpu(self):
         data_file = tu.get_test_data_path('24737.nxs')
         self.test_folder = tempfile.mkdtemp(suffix='my_test/')
         # set options
         options = tu.set_experiment('tomo')
         options['data_file'] = data_file
         options['out_path'] = os.path.join(self.test_folder)
-        options['process_file'] = tu.get_test_process_path('ccpi/ccpi_denoising_gpu_test.nxs')
+        options['process_file'] = tu.get_test_process_path('ccpi/ccpi_denoising_cpu_test.nxs')
         run_protected_plugin_runner(options)
 
         # perform folder cleaning
@@ -51,5 +51,6 @@ class CcpiDenoisingGpuTest(unittest.TestCase):
         os.removedirs(cp_folder)
         classb._empty_folder(self.test_folder)
         os.removedirs(self.test_folder)
+
 if __name__ == "__main__":
     unittest.main()
